[Ide] Allow project's default namespace to be used in templates
The default policy was been changed recently so that files created in
folders take the folder as the namespace. For project templates
that do not want to follow this policy for certain files a new
RootNamespace property is defined which is project's default
namespace and would be the namespace for a file added to the
project directory and not in a sub-folder.